add startstringmatch helper
