+++ libs/designsystem/helpers/src/public_api.ts
@@ -11,6 +11,7 @@ export * from './chart-config-has-type';
 export * from './deep-copy';
 export * from './merge-deep';
 export * from './unique-id-generator.helper';
+export * from './start-string-match.helper';
 export * from './line-clamp-helper';
 export * from './kirby-animation';
 export * from './ionic-element-part-helper';
